requirements.txt
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-ranged-response==0.2.0
django-recaptcha==4.1.0
django-simple-captcha==0.6.2
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
Markdown==3.8.2
MarkupSafe==3.0.2
openai==1.97.1
packaging==25.0
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyOpenSSL==25.1.0
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stripe==12.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
-e git+https://github.com/techwithwayne/agentsuite.git@85738df3d72a834bc7ff430f0404eab49b5fac68#egg=webdoctor
Werkzeug==3.1.3
whitenoise==6.9.0
